requirements.txt
torch
torchvision
gradio
numpy
Pillow
scikit-image
pytorch-lightning